# public/api-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sanduk API
  version: 1.0.0
  description: |
    REST API for Sanduk — a multi-user web app for organizing physical storage bins with QR codes.

    **Authentication**: Most endpoints require a JWT Bearer token obtained via `POST /auth/login`.
    Include it in the `Authorization: Bearer <token>` header.

    **Response envelopes**: All list endpoints return `{ results: T[], count: number }`.
    Error responses use `{ error: "CODE", message: "Human readable" }`.
servers:
  - url: /api
    description: API base path (proxied via nginx)
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from POST /auth/login. Expires after 7 days.
  schemas:
...
    ErrorResponse:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum:
            - VALIDATION_ERROR
            - NOT_FOUND
            - FORBIDDEN
            - CONFLICT
            - RATE_LIMITED
            - INTERNAL_ERROR
        message:
          type: string
      example:
        error: NOT_FOUND
        message: Bin not found

    # ── Auth ───────────────────────────────────────────────
...
    Location:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_by:
          type: string
          format: uuid
          nullable: true
        invite_code:
          type: string
        activity_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 90
          description: Days to keep activity log entries before auto-pruning
        trash_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 30
          description: Days to keep soft-deleted bins before auto-purging
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    LocationWithMeta:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_by:
          type: string
          format: uuid
          nullable: true
        invite_code:
          type: string
        activity_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 90
        trash_retention_days:
          type: integer
          minimum: 7
          maximum: 365
          default: 30
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        member_count:
          type: integer
        role:
          type: string
          enum: [owner, member]
    LocationMember:
      type: object
      properties:
        id:
          type: string
          format: uuid
        location_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, member]
        joined_at:
          type: string
          format: date-time
        display_name:
          type: string
    CreateLocationRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 255
    JoinLocationRequest:
      type: object
      required: [inviteCode]
      properties:
        inviteCode:
          type: string

    # ── Area ───────────────────────────────────────────────
...
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Authentication required
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: FORBIDDEN
            message: Not a member of this location
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: NOT_FOUND
            message: Resource not found
    ValidationError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: VALIDATION_ERROR
            message: Name is required
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: RATE_LIMITED
            message: Too many attempts, please try again later
paths:
...
  /locations:
    get:
      tags: [Locations]
      summary: List user's locations
      description: Returns locations the current user is a member of, with member count and role.
      responses:
        '200':
          description: List of locations
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/LocationWithMeta'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Locations]
      summary: Create a new location
      description: Creates a location and adds the current user as owner.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLocationRequest'
      responses:
        '201':
          description: Location created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
  /locations/join:
    post:
      tags: [Locations]
      summary: Join a location via invite code
      description: Rate limited to 10 per 15 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JoinLocationRequest'
      responses:
        '200':
          description: Joined successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          description: Invalid invite code
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: NOT_FOUND
                message: Invalid invite code
        '409':
          description: Already a member
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                error: CONFLICT
                message: Already a member of this location
        '429':
          $ref: '#/components/responses/RateLimited'
  /locations/{id}:
    put:
      tags: [Locations]
      summary: Update a location
      description: Owner only. Can update name, activity retention, and trash retention.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  maxLength: 255
                activity_retention_days:
                  type: integer
                  minimum: 7
                  maximum: 365
                  description: Days to keep activity log entries
                trash_retention_days:
                  type: integer
                  minimum: 7
                  maximum: 365
                  description: Days to keep soft-deleted bins
      responses:
        '200':
          description: Location updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Location'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Locations]
      summary: Delete a location
      description: Owner only. Cascades to bins, photos, areas, tag colors, and activity log.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Location deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Location deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /locations/{id}/members:
    get:
      tags: [Locations]
      summary: List location members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of members
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/LocationMember'
                  count:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /locations/{id}/members/{userId}:
    delete:
      tags: [Locations]
      summary: Remove a member from location
      description: Owner can remove anyone. Members can only remove themselves (leave).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Member removed
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /locations/{id}/regenerate-invite:
    post:
      tags: [Locations]
      summary: Regenerate invite code
      description: Owner only. Invalidates the previous invite code.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: New invite code
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite_code:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ════════════════════════════════════════════════════════
  # Areas
  # ════════════════════════════════════════════════════════
...
  /shapes/locations:
    get:
      tags: [Shapes]
      summary: Proxy locations shape
      description: Proxies to Electric SQL for real-time location sync for the current user.
      responses:
        '200':
          description: Shape data stream
        '401':
          $ref: '#/components/responses/Unauthorized'
...